requirements.txt
asgiref==3.4.1
cycler==0.10.0
Django==3.2.7
djangorestframework==3.12.4
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.2
Pillow==8.3.2
psycopg2-binary==2.9.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
response==0.5.0
scipy==1.7.1
six==1.16.0
sqlparse==0.4.2
